+++ CHANGELOG.md
@@ -3,7 +3,8 @@ This release introduces supported for [DNS-based](https://www.cloudflare.com/lea
 
 ## New features
 - [PR #979](https://github.com/rqlite/rqlite/pull/979): Add support for DNS-based autoclustering.
-- [PR #985](https://github.com/rqlite/rqlite/pull/985): Add support for DNS SRV-based autoclustering.
+- [PR #979](https://github.com/rqlite/rqlite/pull/979): Add support for DNS-based autoclustering. Fixes [issue #554](https://github.com/rqlite/rqlite/issues/554)
+- [PR #985](https://github.com/rqlite/rqlite/pull/985): Add support for DNS SRV -based autoclustering. Fixes [issue #554](https://github.com/rqlite/rqlite/issues/554)
 
 ### Implementation changes and bug fixes
 - [PR #976](https://github.com/rqlite/rqlite/pull/976): Improve `/readyz` response.