+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 7.3.0 (unreleased)
+## New features
+- [PR #989](https://github.com/rqlite/rqlite/pull/989): Support use of SQLite function `random()`.
+
 ## 7.2.0 (February 1st 2022)
 This release introduces supported for [DNS-based](https://www.cloudflare.com/learning/dns/what-is-dns/) and [DNS SRV -based](https://www.cloudflare.com/learning/dns/dns-records/dns-srv-record/) autoclustering. rqlite can now lookup a DNS record for specified host, and use the network addresses returned to bootstrap the cluster. This can make it much easier to create rqlite clusters using the kinds of Service-Discovery mechanisms you find on Consul and Kubernetes, which are often based on DNS. See the [documentation](https://github.com/rqlite/rqlite/blob/master/DOC/AUTO_CLUSTERING.md) for full details on using DNS-based autoclustering.